requirements.txt
Flask==2.3.3
tweety_ns==0.9.9.1
